+++ connect/server/src/main/scala-spark-master/delta/connect/DeltaPlannerBase.scala
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-package org.apache.spark.sql.delta.connect.server
+package delta.connect
 
+import io.delta.connect.proto
 import io.delta.tables.DeltaTable
 
 import org.apache.spark.sql.SparkSession
-import org.apache.spark.sql.delta.connect.proto
 
 /**
  * Base trait for the planner plugins of Delta Connect.
